+++ app/src/main/java/com/zeapo/pwdstore/UserPreference.kt
@@ -37,7 +37,8 @@ import com.zeapo.pwdstore.autofill.AutofillPreferenceActivity
 import com.zeapo.pwdstore.autofill.oreo.BrowserAutofillSupportLevel
 import com.zeapo.pwdstore.autofill.oreo.getInstalledBrowsersWithAutofillSupportLevel
 import com.zeapo.pwdstore.crypto.PgpActivity
-import com.zeapo.pwdstore.git.GitActivity
+import com.zeapo.pwdstore.git.GitConfigActivity
+import com.zeapo.pwdstore.git.GitServerConfigActivity
 import com.zeapo.pwdstore.pwgenxkpwd.XkpwdDictionary
 import com.zeapo.pwdstore.sshkeygen.ShowSshKeyFragment
 import com.zeapo.pwdstore.sshkeygen.SshKeyGenActivity
@@ -45,6 +46,7 @@ import com.zeapo.pwdstore.utils.PasswordRepository
 import com.zeapo.pwdstore.utils.auth.AuthenticationResult
 import com.zeapo.pwdstore.utils.auth.Authenticator
 import com.zeapo.pwdstore.utils.autofillManager
+import com.zeapo.pwdstore.utils.getEncryptedPrefs
 import java.io.File
 import java.io.IOException
 import java.time.LocalDateTime
@@ -73,6 +75,7 @@ class UserPreference : AppCompatActivity() {
             callingActivity = requireActivity() as UserPreference
             val context = requireContext()
             val sharedPreferences = preferenceManager.sharedPreferences
+            val encryptedPreferences = requireActivity().applicationContext.getEncryptedPrefs("git_operation")
 
             addPreferencesFromResource(R.xml.preference)
 
@@ -121,7 +124,7 @@ class UserPreference : AppCompatActivity() {
             selectExternalGitRepositoryPreference?.summary = sharedPreferences.getString("git_external_repo", getString(R.string.no_repo_selected))
             viewSshKeyPreference?.isVisible = sharedPreferences.getBoolean("use_generated_key", false)
             deleteRepoPreference?.isVisible = !sharedPreferences.getBoolean("git_external", false)
-            sshClearPassphrasePreference?.isVisible = sharedPreferences.getString("ssh_key_passphrase", null)?.isNotEmpty()
+            sshClearPassphrasePreference?.isVisible = encryptedPreferences.getString("ssh_key_local_passphrase", null)?.isNotEmpty()
                     ?: false
             clearHotpIncrementPreference?.isVisible = sharedPreferences.getBoolean("hotp_remember_check", false)
             clearAfterCopyPreference?.isVisible = sharedPreferences.getString("general_show_time", "45")?.toInt() != 0
@@ -172,7 +175,7 @@ class UserPreference : AppCompatActivity() {
             }
 
             sshClearPassphrasePreference?.onPreferenceClickListener = ClickListener {
-                sharedPreferences.edit().putString("ssh_key_passphrase", null).apply()
+                encryptedPreferences.edit().putString("ssh_key_local_passphrase", null).apply()
                 it.isVisible = false
                 true
             }
@@ -190,16 +193,12 @@ class UserPreference : AppCompatActivity() {
             }
 
             gitServerPreference?.onPreferenceClickListener = ClickListener {
-                val intent = Intent(callingActivity, GitActivity::class.java)
-                intent.putExtra("Operation", GitActivity.EDIT_SERVER)
-                startActivityForResult(intent, EDIT_GIT_INFO)
+                startActivity(Intent(callingActivity, GitServerConfigActivity::class.java))
                 true
             }
 
             gitConfigPreference?.onPreferenceClickListener = ClickListener {
-                val intent = Intent(callingActivity, GitActivity::class.java)
-                intent.putExtra("Operation", GitActivity.EDIT_GIT_CONFIG)
-                startActivityForResult(intent, EDIT_GIT_CONFIG)
+                startActivity(Intent(callingActivity, GitConfigActivity::class.java))
                 true
             }