+++ app/src/main/java/com/zeapo/pwdstore/utils/Extensions.kt
@@ -5,10 +5,18 @@
 package com.zeapo.pwdstore.utils
 
 import android.content.Context
+import android.content.SharedPreferences
 import android.os.Build
 import android.util.TypedValue
+import android.view.View
 import android.view.autofill.AutofillManager
+import android.view.inputmethod.InputMethodManager
+import androidx.annotation.IdRes
 import androidx.annotation.RequiresApi
+import androidx.appcompat.app.AlertDialog
+import androidx.core.content.getSystemService
+import androidx.security.crypto.EncryptedSharedPreferences
+import androidx.security.crypto.MasterKeys
 
 infix fun Int.hasFlag(flag: Int): Boolean {
     return this and flag == flag
@@ -22,6 +30,37 @@ fun Context.resolveAttribute(attr: Int): Int {
     val typedValue = TypedValue()
     this.theme.resolveAttribute(attr, typedValue, true)
     return typedValue.data
+}
+
+fun Context.getEncryptedPrefs(fileName: String): SharedPreferences {
+    val keyGenParameterSpec = MasterKeys.AES256_GCM_SPEC
+    val masterKeyAlias = MasterKeys.getOrCreate(keyGenParameterSpec)
+    return EncryptedSharedPreferences.create(
+            fileName,
+            masterKeyAlias,
+            this,
+            EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
+            EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
+    )
+}
+
+/**
+ * Extension function for [AlertDialog] that requests focus for the
+ * view whose id is [id]. Solution based on a StackOverflow
+ * answer: https://stackoverflow.com/a/13056259/297261
+ */
+fun <T : View> AlertDialog.requestInputFocusOnView(@IdRes id: Int) {
+    setOnShowListener {
+        findViewById<T>(id)?.apply {
+            setOnFocusChangeListener { v, _ ->
+                v.post {
+                    context.getSystemService<InputMethodManager>()
+                            ?.showSoftInput(v, InputMethodManager.SHOW_IMPLICIT)
+                }
+            }
+            requestFocus()
+        }
+    }
 }
 
 val Context.autofillManager: AutofillManager?